+++ src/components/product/ProductItem.vue
@@ -12,14 +12,7 @@
       </v-card-title>
     </router-link>
     <v-card-actions class="mb-2">
-      <v-btn v-if="isAdding" @click="addToCart">
-        <v-icon class="mr-1">add_shopping_cart</v-icon>
-        <span class="hidden-md-and-down mt-1">Add to Cart</span>
-      </v-btn>
-      <v-btn v-else @click="removeFromCart(product._id)">
-        <v-icon class="mr-1">remove_shopping_cart</v-icon>
-        <span class="hidden-md-and-down mt-1">Remove from Cart</span>
-      </v-btn>
+      <product-cart-button :product="product"></product-cart-button>
       <v-spacer></v-spacer>
       <span class="subheading hidden-md-and-down">$ {{ product.price }}</span>
       <span class="body-2 hidden-lg-and-up">$ {{ product.price }}</span>
@@ -28,31 +21,12 @@
 </template>
 
 <script>
-  import {
-    ADD_TO_CART,
-    REMOVE_FROM_CART
-  } from '../../store/mutation-types'
-
+  import ProductCartButton from '@/components/product/ProductCartButton'
   export default {
     name: 'product-item',
     props: ['product'],
-    data () {
-      return {
-        cart: this.$store.state.cart
-      }
-    },
-    computed: {
-      isAdding () {
-        return this.cart.indexOf(this.product) < 0
-      }
-    },
-    methods: {
-      addToCart () {
-        this.$store.commit(ADD_TO_CART, this.product)
-      },
-      removeFromCart (id) {
-        this.$store.commit(REMOVE_FROM_CART, id)
-      }
+    components: {
+      'product-cart-button': ProductCartButton
     }
   }
 </script>