+++ src/components/product/ProductList.vue
@@ -14,12 +14,12 @@
     name: 'product-list',
     created () {
       if (this.products.length === 0) {
-        this.$store.dispatch('allProducts')
+        this.$store.dispatch('products')
       }
     },
     computed: {
       products () {
-        return this.$store.getters.allProducts
+        return this.$store.getters.products
       }
     },
     components: {