+++ src/pages/admin/Edit.vue
@@ -1,7 +1,24 @@
 <template>
-  <div>
-    <div class="title">
-      <h1>This is Admin/Edit/{{$route.params.id}}</h1>
-    </div>
-  </div>
+  <product-form @save-product="updateProduct" :product="this.product" :isEditing="true">
+  </product-form>
 </template>
+
+<script>
+  import ProductForm from '@/components/product/ProductForm.vue'
+  export default {
+    props: ['id'],
+    components: {
+      'product-form': ProductForm
+    },
+    computed: {
+      product () {
+        return Object.assign({}, this.$store.getters.product(this.id))
+      }
+    },
+    methods: {
+      updateProduct () {
+        this.$store.dispatch('updateProduct', this.product)
+      }
+    }
+  }
+</script>