+++ CHANGELOG.md
@@ -9,10 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Generated full webpack config w/ vuetifyjs template
 - Separate Header/Footer components
-- Basic navbar with Home, Admin, Cart routes
 - Nested routes under admin page
 - API for accessing a mongodb via mongoose
 - Vuex store for state management, mutation types, getters, etc.
+- Admin page with product catalog, ability to add new or edit current products
+- Details page for products, shared "Add to Cart"/"Remove from Cart" buttons
 
 ### Changed